更新面试查看和报告

+++ src/main/java/com/sysu/smartjob/entity/Interview.java
@@ -33,7 +33,7 @@ public class Interview {
     
     private Integer answeredQuestions;
     
-    private BigDecimal overallScore;
+    private Double overallScore;
     
     private String aiFeedback;
